+++ cli/src/commit_templater.rs
@@ -3161,6 +3161,7 @@ mod tests {
                 &language,
                 &mut diagnostics,
                 text,
+                Default::default(),
                 &self.template_aliases_map,
             )
         }