+++ cli/src/template_builder.rs
@@ -2411,6 +2411,23 @@ fn builtin_functions<'a, L: TemplateLanguage<'a> + ?Sized>() -> TemplateBuildFun
             Ok(L::Property::wrap_template(template))
         },
     );
+    map.insert(
+        "available_width",
+        |_language, _diagnostics, build_ctx, function| {
+            function.expect_no_arguments()?;
+            let width = build_ctx
+                .local_variables
+                .get("$available_width")
+                .ok_or_else(|| {
+                    TemplateParseError::expression("width not available", function.name_span)
+                })?;
+            let width = width()
+                .try_into_integer()
+                .ok()
+                .expect("template language syntax disallows identifiers starting with `$`");
+            Ok(L::Property::wrap_property(width))
+        },
+    );
     map.insert("hash", |language, diagnostics, build_ctx, function| {
         let [content_node] = function.expect_exact_arguments()?;
         let content = expect_stringify_expression(language, diagnostics, build_ctx, content_node)?;
@@ -2811,6 +2828,7 @@ pub fn build_expression<'a, L: TemplateLanguage<'a> + ?Sized>(
 pub fn build<'a, C, L>(
     language: &L,
     diagnostics: &mut TemplateDiagnostics,
+    local_variables: HashMap<&str, &dyn Fn() -> L::Property>,
     node: &ExpressionNode,
 ) -> TemplateParseResult<TemplateRenderer<'a, C>>
 where
@@ -2820,7 +2838,7 @@ where
 {
     let self_placeholder = PropertyPlaceholder::new();
     let build_ctx = BuildContext {
-        local_variables: HashMap::new(),
+        local_variables,
         self_variable: &|| self_placeholder.clone().into_dyn_wrapped(),
     };
     let template = expect_template_expression(language, diagnostics, &build_ctx, node)?;
@@ -2832,6 +2850,7 @@ pub fn parse<'a, C, L>(
     language: &L,
     diagnostics: &mut TemplateDiagnostics,
     template_text: &str,
+    local_variables: HashMap<&str, &dyn Fn() -> L::Property>,
     aliases_map: &TemplateAliasesMap,
 ) -> TemplateParseResult<TemplateRenderer<'a, C>>
 where
@@ -2840,7 +2859,8 @@ where
     L::Property: WrapTemplateProperty<'a, C>,
 {
     let node = template_parser::parse(template_text, aliases_map)?;
-    build(language, diagnostics, &node).map_err(|err| err.extend_alias_candidates(aliases_map))
+    build(language, diagnostics, local_variables, &node)
+        .map_err(|err| err.extend_alias_candidates(aliases_map))
 }
 
 pub fn expect_boolean_expression<'a, L: TemplateLanguage<'a> + ?Sized>(
@@ -3095,6 +3115,7 @@ mod tests {
                 &self.language,
                 &mut TemplateDiagnostics::new(),
                 template,
+                Default::default(),
                 &self.aliases_map,
             )
         }